Add task "mail".
This adds a dedicated task using JavaMail to send mails. In this implementation, the complete mime message will be serialized, encoded as a string, sent trough the AMQP's message stack and deserialized, decoded again. Obviously, this requires an additional setup for a mail account (see internal `config.properties` as a template for additional config props).